+++ source/gear/corrosion-base.js
@@ -28,7 +28,7 @@ module.exports = new GearTemplate("Corrosion",
 		return resultLines;
 	}
 ).setTargetingTags({ type: "single", team: "foe" })
-	.setUpgrades("Fate-Sealing Corrosion", "Harmful Corrosion", "Shattering Corrosion")
+	.setUpgrades("Fate-Sealing Corrosion", "Fatiguing Corrosion", "Shattering Corrosion")
 	.setModifiers({ name: "Power Down", stacks: 20 })
 	.setBonus(2) // Crit Stagger
 	.setCharges(15);